Initializer with non-constant values works with world/global arrays
With the `static` qualifier, a world/global variable is initialized
once and at compile-time. Since at this time there is no way to
initialize a world/global variable at compile-time, this does not work.

Without the `static` qualifier, a world/global variable is initialized
every time it is encountered. Values can be non-constant.