Add compile-time library
When a #library directive lacks a string argument, the library is considered
compile-time only. Such a library can only contain constant, compile-time
objects such as enumerations and builtin function declarations. Such a library
will not be loaded at run-time.

This feature is one way to fix the problem of #including the zcommon.acs file
in multiple libraries. (Since each library is compiled separately, an #include
in each library will be executed and will cause duplication error messages to
be emitted.) By putting the constant objects in a library that only works at
compile-time, the zcommon.acs file will be parsed once.

A compile-time library can also be used to hide constants. By putting constants
in a compile-time library, only the user that #imports the library will see the
constants.