Read `zraw` type
The `zraw` type is used to hold the representation of the other primitive
types. The `zraw` type is a subclass of the `zint` type. `int`, `str`,
and `bool` are aliases for `zraw`.

To stay compatible as much as possible with existing ACS code, the `zraw`
type has special properties. When a `zraw` operand is used with an operand o
a different type, such as in a binary operation, the operand with the differ
type is converted to `zraw`. A value of any other primitive type can be
assigned to a variable of `zraw` type. Similarly, a value of `zraw` type
can be assigned to a variable of any other primitive type.

The `zraw` type should be avoided in new projects.