Fix: inferred raw variables have block scope in weak mode
I thought about making foreach iteration variables have function scope as well,
but I didn't like it in the end. Some problems:
- The user will not be able to reuse the iteration variable names in a
  subsequent foreach loop.
- It takes up more variables in the object file.
- The user can access whatever internal data is stored in the iteration
  variables after the loop has ended, since the variables are still visible.

Yes, the foreach iteration variables are now inconsistent with other variable
declarations scope-wise, but I think it's better to keep it as-is. This is not
a problem in strong mode.