Start variable declaration with `enum`, `struct`, and `synon`
Require the prefix to simplify the parsing. If we allow a variable declaration
to start with an identifier, it will conflict with expressions. For example:

   TestT& t;

The parser cannot tell whether this is a variable declaration or an expression
consisting of a bitwise-and operation.

To solve this problem, another pass can be added to the compiler. This pass
will determine what the identifiers refer to. Then based on that information,
the appropriate constructs will be parsed. At this time, I decided to avoid
this.

This is a well-known problem in parsing of C/C++.