Warn when a script is of the form: script 1 open <statement>
In this form, the body of the script is not a block, but we are skipping a
block when reading an imported script. This will cause the next available block
to be skipped. This means a function could be skipped, because the next block
might be the body of a function:

script 1 open Delay( 123 );
function void f( void ) {} // This block will be skipped, and so will `f`.

This is also a bug in acc.